moved cache type
